TravelPlanner/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
groq==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
plotly==5.17.0
